move landing page to neofab.ai
