+++ public/index.html
@@ -4,11 +4,19 @@
     <meta charset="utf-8" />
     <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#000000" />
+    <meta name="theme-color" content="#203a43" />
     <meta
       name="description"
-      content="Web site created using create-react-app"
+      content="NeoFabAI - The Future of Manufacturing. AI-powered platform that transforms designs into scalable factory layouts, accelerating manufacturing growth for innovators."
     />
+    <meta name="keywords" content="AI manufacturing, factory layout, design automation, manufacturing planning, NeoFab" />
+    <meta property="og:title" content="NeoFabAI - The Future of Manufacturing" />
+    <meta property="og:description" content="AI-powered platform that transforms designs into scalable factory layouts, accelerating manufacturing growth for innovators." />
+    <meta property="og:type" content="website" />
+    <meta property="og:url" content="https://neofab.ai" />
+    <meta name="twitter:card" content="summary_large_image" />
+    <meta name="twitter:title" content="NeoFabAI - The Future of Manufacturing" />
+    <meta name="twitter:description" content="AI-powered platform that transforms designs into scalable factory layouts, accelerating manufacturing growth for innovators." />
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <!--
       manifest.json provides metadata used when your web app is installed on a
@@ -24,7 +32,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>Neofab</title>
+    <title>NeoFabAI - The Future of Manufacturing</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
