--- src/animations/textAnimations.ts
@@ -1,23 +1,0 @@
-export const fadeIn = {
-    initial: { opacity: 0 },
-    animate: { opacity: 1 },
-    transition: { duration: 0.5 }
-  };
-  
-  export const slideInFromLeft = {
-    initial: { x: -100, opacity: 0 },
-    animate: { x: 0, opacity: 1 },
-    transition: { duration: 0.5, type: 'spring' }
-  };
-  
-  export const slideInFromRight = {
-    initial: { x: 100, opacity: 0 },
-    animate: { x: 0, opacity: 1 },
-    transition: { duration: 0.5, type: 'spring' }
-  };
-  
-  export const scaleIn = {
-    initial: { scale: 0, opacity: 0 },
-    animate: { scale: 1, opacity: 1 },
-    transition: { duration: 0.5 }
-  };
