+++ src/assets/emoEnergy.jpeg
