--- src/components/layout/Header.tsx
@@ -1,101 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import { Link } from 'react-router-dom';
-import styled from 'styled-components';
-import { motion } from 'framer-motion';
-import logoImg from '../../assets/images/logo.png';
-
-const HeaderContainer = styled(motion.header)`
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  padding: 1rem 2rem;
-  background-color: rgba(0, 0, 0, 0.8);
-  backdrop-filter: blur(10px);
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  z-index: 1000;
-`;
-
-const Logo = styled(motion.img)`
-  width: 5rem;
-  max-width: 100px; 
-  height: auto;
-  font-family: ${props => props.theme.fonts.headings};
-`;
-
-const Nav = styled.nav`
-  display: flex;
-  gap: 2rem;
-`;
-
-const NavLink = styled(motion(Link))`
-  color: ${props => props.theme.colors.text};
-  font-size: 1.2rem;
-  text-transform: uppercase;
-  letter-spacing: 2px;
-  position: relative;
-
-  &::after {
-    content: '';
-    position: absolute;
-    bottom: -5px;
-    left: 0;
-    width: 0;
-    height: 2px;
-    background-color: ${props => props.theme.colors.primary};
-    transition: width 0.3s ease;
-  }
-
-  &:hover::after {
-    width: 100%;
-  }
-`;
-
-const Header: React.FC = () => {
-  const [scrolled, setScrolled] = useState(false);
-
-  useEffect(() => {
-    const handleScroll = () => {
-      setScrolled(window.scrollY > 50);
-    };
-
-    window.addEventListener('scroll', handleScroll);
-    return () => window.removeEventListener('scroll', handleScroll);
-  }, []);
-
-  return (
-    <HeaderContainer
-      initial={{ y: -100 }}
-      animate={{ y: 0 }}
-      transition={{ duration: 0.5 }}
-      style={{ backgroundColor: scrolled ? 'rgba(0, 0, 0, 0.9)' : 'transparent' }}
-    >
-      <Logo
-        src={logoImg}
-        alt="Logo"
-        initial={{ opacity: 0 }}
-        animate={{ opacity: 1 }}
-        transition={{ delay: 0.2 }}
-      />
-      <Nav>
-        {['Home', 'Product', 'About'].map((item, index) => (
-          <NavLink
-            key={item}
-            to={item === 'Home' ? '/' : `/${item.toLowerCase()}`}
-            initial={{ opacity: 0, y: -20 }}
-            animate={{ opacity: 1, y: 0 }}
-            transition={{ delay: 0.3 + index * 0.1 }}
-            whileHover={{ scale: 1.1 }}
-            whileTap={{ scale: 0.9 }}
-          >
-            {item}
-          </NavLink>
-        ))}
-      </Nav>
-    </HeaderContainer>
-  );
-};
-
-export default Header;
