--- src/styles/theme.ts
@@ -1,24 +1,0 @@
-const theme = {
-  colors: {
-    primary: '#00f0ff',
-    secondary: '#ff00ff',
-    tertiary: '#ffff00',
-    background: '#050505',
-    text: '#ffffff',
-    gradients: {
-      primary: 'linear-gradient(45deg, #00f0ff, #0080ff)',
-      secondary: 'linear-gradient(45deg, #ff00ff, #ff8000)',
-    },
-  },
-  fonts: {
-    main: "'Poppins', sans-serif",
-    headings: "'Syncopate', sans-serif",
-  },
-  breakpoints: {
-    mobile: '576px',
-    tablet: '768px',
-    desktop: '1024px',
-  },
-};
-
-export default theme;
